Added the GlobalExceptionHandler
