+++ src/main/kotlin/nomic/api/RulesAmendments.kt
@@ -1,15 +1,12 @@
 package nomic.api
 
 import nomic.api.models.RulesAmendmentsApiModel
-import nomic.data.EntityNotFoundException
 import nomic.domain.rulesamendments.RuleAmendmentDomain
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
-import org.springframework.web.bind.annotation.ExceptionHandler
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.RequestMapping
-import org.springframework.web.bind.annotation.ResponseStatus
 import org.springframework.web.bind.annotation.RestController
 
 @RestController
@@ -22,29 +19,5 @@ class RulesAmendments(val ruleAmendmentDomain: RuleAmendmentDomain) {
 
         // Return the response object
         return ResponseEntity(rulesAmendments, HttpStatus.OK)
-    }
-
-    @ExceptionHandler(IllegalArgumentException::class)
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    fun handleIllegalArgumentException(
-        exception: IllegalArgumentException
-    ): ResponseEntity<Any> {
-        return ResponseEntity(ExceptionResponseFormat(true, exception.message), HttpStatus.BAD_REQUEST)
-    }
-
-    @ExceptionHandler(EntityNotFoundException::class)
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    fun handleEntityNotFoundException(
-        exception: EntityNotFoundException
-    ): ResponseEntity<Any> {
-        return ResponseEntity(ExceptionResponseFormat(true, exception.message), HttpStatus.NOT_FOUND)
-    }
-
-    @ExceptionHandler(Exception::class)
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    fun handleAllUncaughtException(
-        exception: IllegalArgumentException
-    ): ResponseEntity<Any> {
-        return ResponseEntity(ExceptionResponseFormat(true, "Internal server error"), HttpStatus.INTERNAL_SERVER_ERROR)
     }
 }
